requirements.txt
fastapi
uvicorn
pydantic
google-genai
python-dotenv
google-search-results
sentence-transformers
faiss-cpu
numpy
